requirements-dev.txt
fastapi>=0.110
pyyaml>=6.0
pytest>=7.0
ruff>=0.6
httpx>=0.27.0
